pyproject.toml
[tool.poetry]
name = "home-credit"
version = "0.3.0"
description = ""
authors = ["nastjamakh <nastja.makhonina.dev@gmail.com>"]

[tool.poetry.scripts]
credit = "entrypoint:cli"
train = "train:cli"
data = "data_loader:cli"
model = "modelling.estimator:cli"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
matplotlib = "^3.5.1"
pandas = "^1.4.2"
numpy = ">=1.18,<1.22"
statsmodels = "^0.13.2"
python-dotenv = "^0.20.0"
seaborn = "^0.11.2"
plotly = "^5.6.0"
jupyter = "^1.0.0"
ipython = "^8.2.0"
flake8 = "^4.0.1"
black = "^22.3.0"
wheel = "^0.37.1"
lightgbm = "^3.3.2"
fastapi = "^0.75.1"
requests = "^2.27.1"
uvicorn = "^0.17.6"
missingno = "^0.5.1"
pre-commit = "^2.18.1"
coloredlogs = "^15.0.1"
tqdm = "^4.64.0"
pyarrow = "^7.0.0"
jupyterlab = "^3.3.3"
click = "^8.1.2"
boto3 = "^1.22.8"
botocore = "^1.25.8"
fire = "^0.4.0"
mypy = "^0.950"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
